generate proto; eliminate very-old protobuf dep; re-run mdatagen (edit metadata.yaml) and fix some tests

+++ collector/components/lightstepreceiver/to_traces_test.go
@@ -379,7 +379,7 @@ func TestTimestampOffset(t *testing.T) {
 			},
 		},
 		// Important parameter.
-		TimestampOffsetMicros: int32(offset.Microseconds()),
+		TimestampOffsetMicros: offset.Microseconds(),
 		Spans: []*collectorpb.Span{
 			{
 				OperationName:  "span1",
